+++ src/views/ProjectsSetting.vue
@@ -71,18 +71,32 @@ export default {
   methods: {
     deleteProjects() {
       if (!this.deleteProjectsId.length) return
-      this.$store
-        .dispatch('deleteProjects', this.deleteProjectsId)
+
+      this.$confirm('確定要刪除？', {
+        confirmButtonText: '確定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      })
         .then(() => {
-          this.$message({
-            message: `成功刪除`,
-            type: 'success',
-            center: true,
-            duration: 1800
-          })
+          this.$store
+            .dispatch('deleteProjects', this.deleteProjectsId)
+            .then(() => {
+              this.$message({
+                message: `成功刪除`,
+                type: 'success',
+                center: true,
+                duration: 1800
+              })
+            })
+            .catch(e => {
+              this.$message.error(`請重新檢查 ${e.message}`)
+            })
         })
-        .catch(e => {
-          this.$message.error(`請重新檢查 ${e.message}`)
+        .catch(() => {
+          this.$message({
+            type: 'info',
+            message: '已取消刪除'
+          })
         })
     },
     updateDeleteList(value) {
