+++ src/views/SOItems.vue
@@ -54,19 +54,32 @@ export default {
   methods: {
     deleteSOItems() {
       if (this.deleteList.length === 0) return
-      this.$store
-        .dispatch('deleteSOItems', this.deleteList)
+      this.$confirm('確定要刪除？', {
+        confirmButtonText: '確定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      })
         .then(() => {
-          this.$message({
-            message: `成功刪除`,
-            type: 'success',
-            showClose: true,
-            center: true,
-            duration: 1200
-          })
+          this.$store
+            .dispatch('deleteSOItems', this.deleteList)
+            .then(() => {
+              this.$message({
+                message: `成功刪除`,
+                type: 'success',
+                showClose: true,
+                center: true,
+                duration: 1200
+              })
+            })
+            .catch(() => {
+              this.$message({ message: `已選定 OPT 不能刪除`, type: 'error' })
+            })
         })
         .catch(() => {
-          this.$message({ message: `已選定 OPT 不能刪除`, type: 'error' })
+          this.$message({
+            type: 'info',
+            message: '已取消刪除'
+          })
         })
     },
     updateDeleteList(value) {
